Renaming f-score scripts to remove hyphens, updating to use argparse, simplifying dir script
